requirements.txt
openai-whisper==20240930
ffmpeg-python==0.2.0
torch==2.7.1+cu128